Expose build-platforms for container builds
In order to be able to build aarch64 containers for NVIDIA, we
need Konflux param `build-platforms` to be exposed in the Tekton
template for container images. The default value is `linux/x86_64`
to not modify existing pipelines, while allowing overrides for
specific images.

The main difference with !156 is that it removes the static
`PLATFORM` param in the `build-container` task, so there is no conflict.
It also modifies the `show-sbom` task to show the SBOM for all
platforms.

Signed-off-by: Fabien Dupont <fdupont@redhat.com>