Expose build-platforms for container builds
In order to be able to build aarch64 containers for NVIDIA, we need
Konflux param build-platforms to be exposed in the Tekton template
for container images. The default value is linux/x86_64 to not modify
existing pipelines, while allowing overrides for specific images.

The change is inspired by https://github.com/konflux-ci/olm-operator-konflux-sample/blob/main/.tekton/multi-arch-build-pipeline.yaml.

Signed-off-by: Fabien Dupont <fdupont@redhat.com>