Specify image append platform param
Without this, each build task for each arch tags each archful image with
the same tag. The last one wins and gets to keep the tag - the others
end up being tagless.

You might think this isn't a problem, because the image ref is supplied
to the build-image-index task by digest ref. But - the problem interacts
with quay.io's garbage collection, which deletes image manifests that
have no tag. Our build pipeline races with that garbage collection
mechanism. If we lose, quay delete all of the arches but the tagged one,
and then our build-image-index task fails.

This new append platform param was itnroduced here:
https://github.com/konflux-ci/build-definitions/pull/2757